double 사용 불가
double을 number(38,2)로 수정
정수부분 38자리까지
소수부분 2자리까지 저장되는 데이터형식

+++ scott1.sql
@@ -2,7 +2,7 @@ create table users
 (
 user_id varchar2(100) primary key,
 user_email varchar2(100) not null unique,
-user_balance double default 0,
+user_balance number(38,2) default 0,
 user_country varchar2(100),
 user_city varchar2(100),   -- 시(ex 대구)
 user_street varchar2(100), -- 도로명 주소
